The last page build is hanging
