+++ src/app/blog/tag/[tag_id]/page.tsx
@@ -4,13 +4,31 @@ import { TagIcon } from "@/components/icons/TagIcon/TagIcon";
 import { PageWrapper } from "@/components/organisms/PageWrapper/PageWrapper";
 import { PostList } from "@/components/organisms/PostList/PostList";
 import { Badge } from "@/components/ui/badge";
-import { getPostsByTag } from "@/utils/blog/posts";
+import { getPosts, getPostsByTag } from "@/utils/blog/posts";
 
 type PageParams = {
   params: {
     tag_id: string;
   };
 };
+
+export async function generateStaticParams(): Promise<PageParams["params"][]> {
+  const allTags = new Set<string>();
+  const allPosts = await getPosts();
+
+  // Get all distinct blog post tags.
+  allPosts.forEach((post) => {
+    post.frontMatter.tags.forEach((tag) => {
+      allTags.add(tag);
+    });
+  });
+
+  return Array.from(allTags).map((tag) => {
+    return {
+      tag_id: tag,
+    };
+  });
+}
 
 const BlogTagPage: NextPage<PageParams> = async ({ params: { tag_id } }) => {
   const posts = await getPostsByTag(tag_id);
