+++ src/utils/photos/getSuggestedPhotoTags.test.ts
@@ -1,5 +1,5 @@
 import { getSuggestedPhotoTags } from "./getSuggestedPhotoTags";
-import { PhotoTags } from "@/constants/photoTags";
+import { PhotoTags } from "@/constants/photoTags/photoTags";
 
 test("get suggested photo tags for a given photo tag", () => {
   const suggestions = getSuggestedPhotoTags(PhotoTags.City, 3);
